requirements.txt
# 核心依赖
mitsuba>=3.0.0
numpy>=1.20.0
pyyaml>=5.4.0

# 可选依赖（推荐安装）
opencv-python>=4.5.0  # 用于图像格式转换
matplotlib>=3.3.0     # 用于数据集可视化
